# apis/product/productApi.js
'use strict'
const express = require('express');
const MysqlDB = require('../../models/mysql');
const ProductService = require('../../services/productService/productService');
const productApi = express.Router();
const mysqlDb = new MysqlDB();
const productService = new ProductService(mysqlDb);
const {checkRequiredFieldInBody,checkRequiredFieldInQuery} = require('../../middleware/index')
const {verifyToken,adminRole} = require('../../middleware/verifyToken');
const {removeAccent} =  require('../../utils/index');
productApi.get('/', (req,res,next) => {
    let {minPrice,maxPrice,productsPerPage,pageNumber,orderType,search} = req.query;
    productService
    .getProducts(minPrice,maxPrice,productsPerPage,pageNumber,orderType,search)
    .then(listProduct => {
        return res.status(200).json({status:200,message:"Success",data: listProduct})
    })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })

})
productApi.get('/get-by-category-id/:category_id', (req,res,next) => {
    
    let {category_id} = req.params
    let {productsPerPage,pageNumber,orderType,search} = req.query;
    productService
    .getProductsByCategoryId(category_id,productsPerPage,pageNumber,orderType,search)
    .then(listProduct => {
        return res.status(200).json({status:200,message:"Success",data: listProduct})
    })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })

})
productApi.get('/get-by-category-slug/:category_slug', (req,res,next) => {
    
    let {category_slug} = req.params

    let {minPrice,maxPrice,productsPerPage,pageNumber,orderType,search} = req.query;
    productService
    .getProductsByCategorySlug(category_slug,minPrice,maxPrice,productsPerPage,pageNumber,orderType,search)
    .then(listProduct => {
        return res.status(200).json({status:200,message:"Success",data: listProduct})
    })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })

})

productApi.get('/:id',(req,res,next)=>{
    let {id} = req.params
    productService
    .getProductById(id)
    .then(listProduct=>{
        return res.status(200).json({status:200,message:"Success",data: listProduct})
        })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })  
})
productApi.get('/get-by-slug/:slug',(req,res,next)=>{
    let {slug} = req.params
    productService
    .getProductBySlug(slug)
    .then(listProduct=>{
        return res.status(200).json({status:200,message:"Success",data: listProduct})
        })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })  
})
productApi.post('/',verifyToken,adminRole,checkRequiredFieldInBody(['name','category_id']), (req,res,next)=>{
    let {name,description,detail,list_product_images,price,discount,category_id} = req.body
    productService
    .createProduct(name, description, detail,list_product_images, price, discount, category_id)
    .then(result => { 
        return res.status(200).json({
            status:200,
            message: 'Post new product successfully'
        })
    })
    .catch(err => {
        return res.status(500).json({status:500,message: err})
    })  
})
productApi.put('/:id',verifyToken,adminRole, (req,res,next)=>{
    let {id} = req.params
    let {name,description,detail,list_product_images,price,discount,category_id} = req.body
    productService
    .updateProduct(id,name,description,detail,list_product_images,price,discount,category_id)
    .then(result=>{
        return res.status(200).json({  
            status:200,
            message: "Update product successfully",
            data:result
            })
        })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })  
})
productApi.delete('/:id',verifyToken,adminRole,(req,res,next)=>{
    let {id} = req.params
    productService
    .deleteProduct(id)
    .then(result=>{
        return res.status(200).json({
            status:200,
            message: 'Detele product sucessfully',
            })
    })
    .catch(err=>{
        return res.status(500).json({status:500,message: err})
    })  
})

module.exports = productApi;