# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Camera Apis",
    "version": "1.0.0",
    "title": "Camera Apis",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:5000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/product/": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "minPrice",
            "type": "integer",
            "in": "query",
            "description": "MinPrice",
            "required": false
          },
          {
            "name": "maxPrice",
            "in": "query",
            "type":"integer",
            "description": "Max Price ",
            "required": false
          },
          {
            "name": "productsPerPage",
            "in": "query",
            "description": "Max products can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 1-oldest 2-newest ",
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "for search request for Front End",
            "required": false
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "require": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "detail": {
                  "type": "string"
                },
                "price": {
                  "type": "integer"
                },
                "discount": {
                  "type": "integer"
                },
                "list_product_images": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minItems": 0
                  }
                },
                "category_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Create Product",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                },
                "insertedId": {
                  "type": "integer"
                }
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/get-by-category-id/{category_id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category_id",
            "type": "integer",
            "in": "path",
            "required": true
          },
          {
            "name": "productsPerPage",
            "in": "query",
            "description": "Max products can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 1-oldest 2-newest ",
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "for search request for Front End",
            "required": false
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products By Category",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/get-by-category-slug/{category_slug}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "category_slug",
            "type": "string",
            "in": "path",
            "description": "category slug",
            "required": true
          },
          {
            "name": "minPrice",
            "in": "query",
            "type":"integer",
            "description": "Min Price ",
            "required": false
          },
          { 
            "name": "maxPrice",
            "in": "query",
            "type":"integer",
            "description": "Max Price ",
            "required": false
          },
          {
            "name": "productsPerPage",
            "in": "query",
            "description": "Max products can be display in one page",
            "required": false
          },
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Page number",
            "required": false
          },
          {
            "name": "orderType",
            "in": "query",
            "description": "Order type to display: 1-oldest 2-newest ",
            "required": false
          },
          {
            "name": "search",
            "in": "query",
            "description": "for search request for Front End",
            "required": false
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products By Category Slug",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/get-by-slug/{slug}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "slug",
            "type": "string",
            "in": "path",
            "description": "slug of product",
            "required": true
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Products By Slug",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/product"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
    "/product/{id}": {
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Get Product by Id",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/product"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "put": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "description": "ID of post",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "detail": {
                  "type": "string"
                },
                "price": {
                  "type": "integer"
                },
                "discount": {
                  "type": "integer"
                },
                "category_id": {
                  "type": "integer"
                },
                "list_product_images":{
                  "type": "array",
                  "items": {
                    "type": "string",
                    "minItems": 0
                  }
                }
              }
            }
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Update Product by ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-access-token",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "type": "integer",
            "in": "path",
            "required": true
          }
        ],
        "tags": [
          "Product Api"
        ],
        "summary": "Delete product by ID",
        "description": "",
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "$ref": "#/definitions/success"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "detail":{
          "type": "string"
        },
        "list_product_images": {
          "type": "array",
          "items": {
            "type": "string",
            "minItems": 0
          }
        },
        "price": {
          "type": "integer"
        },
        "discount": {
          "type": "integer"
        },
        "category_id": {
          "$ref": "#/definitions/category"
        },
        "slug": {
          "type": "string"
        },
        "created_at": {
          "type": "string",
          "example": "2020-01-08T13:25:50"
        },
        "updated_at": {
          "type": "string",
          "example": "2020-01-08T13:25:50"
        }
      }
    },
...
    "category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "slug": {
          "type": "string"
        }
      }
    },
...
    "error": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "success": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}